% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rsd_et_al.r
\name{gcv}
\alias{gcv}
\title{Geometric Coefficient of Variation}
\usage{
gcv(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
This function calculates the geometric coefficient of variation (GCV) of
a numeric vector.
}
\examples{
gcv(c(19, 17, 18))
}

