% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rsd_et_al.r
\name{geomean}
\alias{geomean}
\title{Geometric Mean}
\usage{
geomean(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
This function calculates the geometric mean of a numeric vector.
}
\examples{
geomean(c(19, 17, 18))
}

