% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_fns.r
\name{plot_fn}
\alias{plot_fn}
\title{Plot Function}
\usage{
plot_fn(f, start, end, res = 250, ...)
}
\arguments{
\item{f}{a function}

\item{start}{value at which to begin plotting}

\item{end}{value at which to stop plotting}

\item{res}{resolution of the plot}

\item{...}{additional arguments to be passed to \code{f}}
}
\description{
This function plots a numeric function that takes a numeric input.
}
\examples{
plot_fn(sin, -pi, pi)
plot_fn(dgamma, 0, 5, shape = 1.1, rate = 2)
}

