% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_fns.r
\name{theme_galangjs}
\alias{theme_galangjs}
\title{Joel's ggplot Theme}
\usage{
theme_galangjs()
}
\description{
A ggplot plot theme adds the the Candara font to \code{theme_solarize} from
the \code{ggthemes} package.
}
\examples{
plot_fn(sin, -pi, pi) + theme_galangjs()
}

