% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rsd_et_al.r
\name{gcv}
\alias{gcv}
\title{Geometric Coefficient of Variation}
\usage{
gcv(x, as_pct = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{as_pct}{logical stating whether the result should be expressed as a
percentage}
}
\description{
This function calculates the geometric coefficient of variation (GCV) of
a numeric vector.
}
\examples{
gcv(c(19, 17, 18))
gcv(17:19, as_pct = TRUE)
}

