% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_fns.r
\name{theme_galangjs}
\alias{theme_galangjs}
\title{Joel's ggplot Theme}
\usage{
theme_galangjs()
}
\description{
A ggplot plot theme based on \code{theme_bw} that uses the Candara font and
a solarized light background.
}
\examples{
plot_fn(sin, -pi, pi) + theme_galangjs()
}

